add `_utils_essentials` to avoid circular imports
