+++ AlbertoX3/environment.py
@@ -22,7 +22,7 @@ __all__ = (
 
 from dotenv import load_dotenv
 from os import environ, getenv
-from .utils import get_bool
+from ._utils_essentials import get_bool
 
 
 load_dotenv()
